requirement.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-debug-toolbar==1.11
django-redis==4.10.0
django-shortuuidfield==0.1.3
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
idna==2.8
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2018.9
qiniu==7.2.2
redis==3.2.1
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
